@@ -1,9 +1,12 @@
+from __future__ import annotations
+
 import inspect
 import numbers
 import operator
+import typing
 import warnings
 from functools import lru_cache
-from typing import ClassVar
+from typing import Any, ClassVar, Literal
 
 import numpy as np
 import pandas as pd
@@ -23,17 +26,25 @@ from shapely.geometry.base import BaseGeometry
 from ._compat import (
     GEOS_GE_312,
     HAS_PYPROJ,
-    PANDAS_GE_21,
-    PANDAS_GE_22,
-    SHAPELY_GE_21,
     requires_pyproj,
 )
 from .sindex import SpatialIndex
+
+if typing.TYPE_CHECKING:
+    import numpy.typing as npt
+
+    from .base import GeoPandasBase
+
+    if HAS_PYPROJ:
+        from pyproj import CRS
+    else:
+        CRS = Any
 
 if HAS_PYPROJ:
     from pyproj import Transformer
 
     TransformerFromCRS = lru_cache(Transformer.from_crs)
+
 
 _names = {
     "MISSING": None,
@@ -81,7 +92,9 @@ class GeometryDtype(ExtensionDtype):
 register_extension_dtype(GeometryDtype)
 
 
-def _check_crs(left, right, allow_none=False):
+def _check_crs(
+    left: GeoPandasBase, right: GeoPandasBase, allow_none: bool = False
+) -> bool:
     """
     Check if the projection of both arrays is the same.
 
@@ -95,7 +108,9 @@ def _check_crs(left, right, allow_none=False):
     return True
 
 
-def _crs_mismatch_warn(left, right, stacklevel=3):
+def _crs_mismatch_warn(
+    left: GeoPandasBase, right: GeoPandasBase, stacklevel: int = 3
+) -> None:
     """Raise a CRS mismatch warning with the information on the assigned CRS."""
     if left.crs:
         left_srs = left.crs.to_string()
@@ -123,7 +138,7 @@ def _crs_mismatch_warn(left, right, stacklevel=3):
     )
 
 
-def isna(value):
+def isna(value: None | float | pd.NA) -> bool:
     """
     Check if scalar value is NA-like (None, np.nan or pd.NA).
 
@@ -145,11 +160,11 @@ def isna(value):
 # -----------------------------------------------------------------------------
 
 
-def _is_scalar_geometry(geom):
+def _is_scalar_geometry(geom) -> bool:
     return isinstance(geom, BaseGeometry)
 
 
-def from_shapely(data, crs=None):
+def from_shapely(data, crs: CRS | None = None) -> GeometryArray:
     """
     Convert a list or array of shapely objects to a GeometryArray.
 
@@ -191,14 +206,18 @@ def from_shapely(data, crs=None):
     return GeometryArray(arr, crs=crs)
 
 
-def to_shapely(geoms):
+def to_shapely(geoms: GeometryArray) -> np.ndarray:
     """Convert GeometryArray to numpy object array of shapely objects."""
     if not isinstance(geoms, GeometryArray):
         raise ValueError("'geoms' must be a GeometryArray")
     return geoms._data
 
 
-def from_wkb(data, crs=None, on_invalid="raise"):
+def from_wkb(
+    data,
+    crs: Any | None = None,
+    on_invalid: Literal["raise", "warn", "ignore"] = "raise",
+) -> GeometryArray:
     """
     Convert a list or array of WKB objects to a GeometryArray.
 
@@ -223,14 +242,18 @@ def from_wkb(data, crs=None, on_invalid="raise"):
     return GeometryArray(shapely.from_wkb(data, on_invalid=on_invalid), crs=crs)
 
 
-def to_wkb(geoms, hex=False, **kwargs):
+def to_wkb(geoms: GeometryArray, hex: bool = False, **kwargs):
     """Convert GeometryArray to a numpy object array of WKB objects."""
     if not isinstance(geoms, GeometryArray):
         raise ValueError("'geoms' must be a GeometryArray")
     return shapely.to_wkb(geoms, hex=hex, **kwargs)
 
 
-def from_wkt(data, crs=None, on_invalid="raise"):
+def from_wkt(
+    data,
+    crs: Any | None = None,
+    on_invalid: Literal["raise", "warn", "ignore"] = "raise",
+) -> GeometryArray:
     """
     Convert a list or array of WKT objects to a GeometryArray.
 
@@ -255,14 +278,16 @@ def from_wkt(data, crs=None, on_invalid="raise"):
     return GeometryArray(shapely.from_wkt(data, on_invalid=on_invalid), crs=crs)
 
 
-def to_wkt(geoms, **kwargs):
+def to_wkt(geoms: GeometryArray, **kwargs):
     """Convert GeometryArray to a numpy object array of WKT objects."""
     if not isinstance(geoms, GeometryArray):
         raise ValueError("'geoms' must be a GeometryArray")
     return shapely.to_wkt(geoms, **kwargs)
 
 
-def points_from_xy(x, y, z=None, crs=None):
+def points_from_xy(
+    x: npt.ArrayLike, y: npt.ArrayLike, z: npt.ArrayLike = None, crs: Any | None = None
+) -> GeometryArray:
     """
     Generate GeometryArray of shapely Point geometries from x, y(, z) coordinates.
 
@@ -321,7 +346,7 @@ class GeometryArray(ExtensionArray):
 
     _dtype = GeometryDtype()
 
-    def __init__(self, data, crs=None):
+    def __init__(self, data, crs: Any | None = None):
         if isinstance(data, self.__class__):
             if not crs:
                 crs = data.crs
@@ -342,14 +367,14 @@ class GeometryArray(ExtensionArray):
         self._sindex = None
 
     @property
-    def sindex(self):
+    def sindex(self) -> SpatialIndex:
         """Spatial index for the geometries in this array."""
         if self._sindex is None:
             self._sindex = SpatialIndex(self._data)
         return self._sindex
 
     @property
-    def has_sindex(self):
+    def has_sindex(self) -> bool:
         """Check the existence of the spatial index without generating it.
 
         Use the `.sindex` attribute on a GeoDataFrame or GeoSeries
@@ -373,7 +398,7 @@ class GeometryArray(ExtensionArray):
         return self._sindex is not None
 
     @property
-    def crs(self):
+    def crs(self) -> CRS:
         """The Coordinate Reference System (CRS) represented as a ``pyproj.CRS`` object.
 
         Returns a ``pyproj.CRS`` or None. When setting, the value
@@ -384,7 +409,7 @@ class GeometryArray(ExtensionArray):
         return self._crs
 
     @crs.setter
-    def crs(self, value):
+    def crs(self, value: Any) -> None:
         """Set the value of the crs."""
         if HAS_PYPROJ:
             from pyproj import CRS
@@ -402,7 +427,7 @@ class GeometryArray(ExtensionArray):
                 )
             self._crs = None
 
-    def check_geographic_crs(self, stacklevel):
+    def check_geographic_crs(self, stacklevel: int) -> None:
         """Check CRS and warn if the planar operation is done in a geographic CRS."""
         if self.crs and self.crs.is_geographic:
             warnings.warn(
@@ -415,13 +440,13 @@ class GeometryArray(ExtensionArray):
             )
 
     @property
-    def dtype(self):
+    def dtype(self) -> GeometryDtype:
         return self._dtype
 
-    def __len__(self):
+    def __len__(self) -> int:
         return self.shape[0]
 
-    def __getitem__(self, idx):
+    def __getitem__(self, idx) -> GeometryArray:
         if isinstance(idx, numbers.Integral):
             return self._data[idx]
         # array-like, slice
@@ -504,17 +529,15 @@ class GeometryArray(ExtensionArray):
     def is_valid_reason(self):
         return shapely.is_valid_reason(self._data)
 
-    def is_valid_coverage(self, gap_width=0.0):
-        if not (SHAPELY_GE_21 and GEOS_GE_312):
-            raise ImportError(
-                "Method 'is_valid_coverage' requires shapely>=2.1 and GEOS>=3.12."
-            )
+    def is_valid_coverage(self, gap_width: float = 0.0):
+        if not GEOS_GE_312:
+            raise ImportError("Method 'is_valid_coverage' requires GEOS>=3.12.")
         return bool(shapely.coverage_is_valid(self._data, gap_width=gap_width))
 
-    def invalid_coverage_edges(self, gap_width=0.0):
-        if not (SHAPELY_GE_21 and GEOS_GE_312):
+    def invalid_coverage_edges(self, gap_width: float = 0.0):
+        if not GEOS_GE_312:
             raise ImportError(
-                "Method 'invalid_coverage_edges' requires shapely>=2.1 and GEOS>=3.12."
+                "Method 'invalid_coverage_edges' requires and GEOS>=3.12."
             )
         return shapely.coverage_invalid_edges(self._data, gap_width=gap_width)
 
@@ -544,9 +567,6 @@ class GeometryArray(ExtensionArray):
 
     @property
     def has_m(self):
-        if not SHAPELY_GE_21:
-            raise ImportError("'has_m' requires shapely>=2.1.")
-
         return shapely.has_m(self._data)
 
     @property
@@ -596,32 +616,29 @@ class GeometryArray(ExtensionArray):
     #
 
     @property
-    def boundary(self):
+    def boundary(self) -> GeometryArray:
         return GeometryArray(shapely.boundary(self._data), crs=self.crs)
 
     @property
-    def centroid(self):
+    def centroid(self) -> GeometryArray:
         self.check_geographic_crs(stacklevel=5)
         return GeometryArray(shapely.centroid(self._data), crs=self.crs)
 
     def concave_hull(self, ratio, allow_holes):
         return shapely.concave_hull(self._data, ratio=ratio, allow_holes=allow_holes)
 
-    def constrained_delaunay_triangles(self):
-        if not SHAPELY_GE_21:
-            raise ImportError("'constrained_delaunay_triangles' requires shapely>=2.1.")
-
+    def constrained_delaunay_triangles(self) -> GeometryArray:
         return GeometryArray(
             shapely.constrained_delaunay_triangles(self._data), crs=self.crs
         )
 
     @property
-    def convex_hull(self):
+    def convex_hull(self) -> GeometryArray:
         """Return the convex hull of the geometries in this array."""
         return GeometryArray(shapely.convex_hull(self._data), crs=self.crs)
 
     @property
-    def envelope(self):
+    def envelope(self) -> GeometryArray:
         """Return the envelope of the geometries in this array."""
         return GeometryArray(shapely.envelope(self._data), crs=self.crs)
 
@@ -630,10 +647,10 @@ class GeometryArray(ExtensionArray):
         return GeometryArray(shapely.oriented_envelope(self._data), crs=self.crs)
 
     @property
-    def exterior(self):
+    def exterior(self) -> GeometryArray:
         return GeometryArray(shapely.get_exterior_ring(self._data), crs=self.crs)
 
-    def extract_unique_points(self):
+    def extract_unique_points(self) -> GeometryArray:
         return GeometryArray(shapely.extract_unique_points(self._data), crs=self.crs)
 
     def offset_curve(self, distance, quad_segs=8, join_style="round", mitre_limit=5.0):
@@ -649,7 +666,7 @@ class GeometryArray(ExtensionArray):
         )
 
     @property
-    def interiors(self):
+    def interiors(self) -> np.ndarray:
         # no GeometryArray as result
         has_non_poly = False
         inner_rings = []
@@ -673,22 +690,19 @@ class GeometryArray(ExtensionArray):
         data[:] = inner_rings
         return data
 
-    def remove_repeated_points(self, tolerance=0.0):
+    def remove_repeated_points(self, tolerance=0.0) -> GeometryArray:
         return GeometryArray(
             shapely.remove_repeated_points(self._data, tolerance=tolerance),
             crs=self.crs,
         )
 
-    def representative_point(self):
+    def representative_point(self) -> GeometryArray:
         return GeometryArray(shapely.point_on_surface(self._data), crs=self.crs)
 
-    def minimum_bounding_circle(self):
+    def minimum_bounding_circle(self) -> GeometryArray:
         return GeometryArray(shapely.minimum_bounding_circle(self._data), crs=self.crs)
 
-    def maximum_inscribed_circle(self, tolerance):
-        if not SHAPELY_GE_21:
-            raise ImportError("'maximum_inscribed_circle' requires shapely>=2.1.")
-
+    def maximum_inscribed_circle(self, tolerance) -> GeometryArray:
         return GeometryArray(
             shapely.maximum_inscribed_circle(self._data, tolerance=tolerance),
             crs=self.crs,
@@ -700,61 +714,56 @@ class GeometryArray(ExtensionArray):
     def minimum_clearance(self):
         return shapely.minimum_clearance(self._data)
 
-    def minimum_clearance_line(self):
-        if not SHAPELY_GE_21:
-            raise ImportError("'minimum_clearance_line' requires shapely>=2.1.")
+    def minimum_clearance_line(self) -> GeometryArray:
         return GeometryArray(shapely.minimum_clearance_line(self._data), crs=self.crs)
 
-    def normalize(self):
+    def normalize(self) -> GeometryArray:
         return GeometryArray(shapely.normalize(self._data), crs=self.crs)
 
-    def orient_polygons(self, exterior_cw=False):
-        if not SHAPELY_GE_21:
-            raise ImportError("'orient_polygons' requires shapely>=2.1.")
+    def orient_polygons(self, exterior_cw: bool = False) -> GeometryArray:
         return GeometryArray(
             shapely.orient_polygons(self._data, exterior_cw=exterior_cw), crs=self.crs
         )
 
-    def make_valid(self, method="linework", keep_collapsed=True):
-        kwargs = {}
-        if SHAPELY_GE_21:
-            kwargs["method"] = method
-            kwargs["keep_collapsed"] = keep_collapsed
-        else:
-            if method != "linework":
-                raise ValueError(
-                    "Only the 'linework' method is supported for shapely < 2.1."
-                )
+    def make_valid(
+        self,
+        method: Literal["linework", "structure"] = "linework",
+        keep_collapsed: bool = True,
+    ) -> GeometryArray:
+        return GeometryArray(
+            shapely.make_valid(
+                self._data, method=method, keep_collapsed=keep_collapsed
+            ),
+            crs=self.crs,
+        )
 
-        return GeometryArray(shapely.make_valid(self._data, **kwargs), crs=self.crs)
-
-    def reverse(self):
+    def reverse(self) -> GeometryArray:
         return GeometryArray(shapely.reverse(self._data), crs=self.crs)
 
-    def segmentize(self, max_segment_length):
+    def segmentize(self, max_segment_length) -> GeometryArray:
         return GeometryArray(
             shapely.segmentize(self._data, max_segment_length),
             crs=self.crs,
         )
 
-    def force_2d(self):
+    def force_2d(self) -> GeometryArray:
         return GeometryArray(shapely.force_2d(self._data), crs=self.crs)
 
-    def force_3d(self, z=0):
+    def force_3d(self, z=0) -> GeometryArray:
         return GeometryArray(shapely.force_3d(self._data, z=z), crs=self.crs)
 
-    def transform(self, transformation, include_z=False):
+    def transform(self, transformation, include_z: bool = False) -> GeometryArray:
         return GeometryArray(
             shapely.transform(self._data, transformation, include_z=include_z),
             crs=self.crs,
         )
 
-    def line_merge(self, directed=False):
+    def line_merge(self, directed: bool = False) -> GeometryArray:
         return GeometryArray(
             shapely.line_merge(self._data, directed=directed), crs=self.crs
         )
 
-    def set_precision(self, grid_size, mode="valid_output"):
+    def set_precision(self, grid_size: float, mode="valid_output"):
         return GeometryArray(
             shapely.set_precision(self._data, grid_size=grid_size, mode=mode),
             crs=self.crs,
@@ -820,32 +829,30 @@ class GeometryArray(ExtensionArray):
         return self._binary_method("equals_exact", self, other, tolerance=tolerance)
 
     def geom_equals_identical(self, other):
-        if not SHAPELY_GE_21:
-            raise ImportError("'geom_equals_identical' requires shapely>=2.1.")
         return self._binary_method("equals_identical", self, other)
 
     #
     # Binary operations that return new geometries
     #
 
-    def clip_by_rect(self, xmin, ymin, xmax, ymax):
+    def clip_by_rect(self, xmin, ymin, xmax, ymax) -> GeometryArray:
         return GeometryArray(
             shapely.clip_by_rect(self._data, xmin, ymin, xmax, ymax), crs=self.crs
         )
 
-    def difference(self, other, grid_size=None):
+    def difference(self, other, grid_size=None) -> GeometryArray:
         return GeometryArray(
             self._binary_method("difference", self, other, grid_size=grid_size),
             crs=self.crs,
         )
 
-    def intersection(self, other, grid_size=None):
+    def intersection(self, other, grid_size=None) -> GeometryArray:
         return GeometryArray(
             self._binary_method("intersection", self, other, grid_size=grid_size),
             crs=self.crs,
         )
 
-    def symmetric_difference(self, other, grid_size=None):
+    def symmetric_difference(self, other, grid_size=None) -> GeometryArray:
         return GeometryArray(
             self._binary_method(
                 "symmetric_difference", self, other, grid_size=grid_size
@@ -853,22 +860,22 @@ class GeometryArray(ExtensionArray):
             crs=self.crs,
         )
 
-    def union(self, other, grid_size=None):
+    def union(self, other, grid_size=None) -> GeometryArray:
         return GeometryArray(
             self._binary_method("union", self, other, grid_size=grid_size), crs=self.crs
         )
 
-    def shortest_line(self, other):
+    def shortest_line(self, other) -> GeometryArray:
         return GeometryArray(
             self._binary_method("shortest_line", self, other), crs=self.crs
         )
 
-    def snap(self, other, tolerance):
+    def snap(self, other, tolerance) -> GeometryArray:
         return GeometryArray(
             self._binary_method("snap", self, other, tolerance=tolerance), crs=self.crs
         )
 
-    def shared_paths(self, other):
+    def shared_paths(self, other) -> GeometryArray:
         return GeometryArray(
             self._binary_method("shared_paths", self, other), crs=self.crs
         )
@@ -889,22 +896,43 @@ class GeometryArray(ExtensionArray):
         self.check_geographic_crs(stacklevel=6)
         return self._binary_method("frechet_distance", self, other, **kwargs)
 
-    def buffer(self, distance, resolution=16, **kwargs):
+    def buffer(self, distance, quad_segs: int | None = None, **kwargs):
         if not (isinstance(distance, int | float) and distance == 0):
             self.check_geographic_crs(stacklevel=5)
+        if "resolution" in kwargs:
+            if quad_segs is not None:
+                msg = (
+                    "`buffer` received both `quad_segs` and `resolution` but these are "
+                    "aliases for the same parameter. Use `quad_segs` only instead."
+                )
+                raise ValueError(msg)
+
+            msg = (
+                "The `resolution` argument to `buffer` is deprecated, `quad_segs` "
+                "should be used instead to align with shapely."
+            )
+            warnings.warn(
+                msg,
+                category=DeprecationWarning,
+                stacklevel=4,
+            )
+            quad_segs = kwargs.pop("resolution")
+        if quad_segs is None:
+            quad_segs = 16  # note shapely default is 8, 16 is historical choice
+
         return GeometryArray(
-            shapely.buffer(self._data, distance, quad_segs=resolution, **kwargs),
+            shapely.buffer(self._data, distance, quad_segs=quad_segs, **kwargs),
             crs=self.crs,
         )
 
-    def interpolate(self, distance, normalized=False):
+    def interpolate(self, distance, normalized: bool = False) -> GeometryArray:
         self.check_geographic_crs(stacklevel=5)
         return GeometryArray(
             shapely.line_interpolate_point(self._data, distance, normalized=normalized),
             crs=self.crs,
         )
 
-    def simplify(self, tolerance, preserve_topology=True):
+    def simplify(self, tolerance, preserve_topology: bool = True) -> GeometryArray:
         return GeometryArray(
             shapely.simplify(
                 self._data, tolerance, preserve_topology=preserve_topology
@@ -912,8 +940,10 @@ class GeometryArray(ExtensionArray):
             crs=self.crs,
         )
 
-    def simplify_coverage(self, tolerance, simplify_boundary=True):
-        if not (SHAPELY_GE_21 and GEOS_GE_312):
+    def simplify_coverage(
+        self, tolerance, simplify_boundary: bool = True
+    ) -> GeometryArray:
+        if not GEOS_GE_312:
             raise ImportError(
                 "'simplify_coverage' requires shapely>=2.1 and GEOS>=3.12."
             )
@@ -960,7 +990,7 @@ class GeometryArray(ExtensionArray):
         elif method == "unary":
             return shapely.union_all(self._data, grid_size=grid_size)
         elif method == "disjoint_subset":
-            if not (SHAPELY_GE_21 and GEOS_GE_312):
+            if not GEOS_GE_312:
                 raise ImportError(
                     "Method 'disjoin_subset' requires shapely>=2.1 and GEOS>=3.12."
                 )
@@ -995,19 +1025,23 @@ class GeometryArray(ExtensionArray):
         data[:] = out
         return data
 
-    def affine_transform(self, matrix):
+    def affine_transform(self, matrix) -> GeometryArray:
         return GeometryArray(
             self._affinity_method("affine_transform", self._data, matrix),
             crs=self.crs,
         )
 
-    def translate(self, xoff=0.0, yoff=0.0, zoff=0.0):
+    def translate(
+        self, xoff: float = 0.0, yoff: float = 0.0, zoff: float = 0.0
+    ) -> GeometryArray:
         return GeometryArray(
             self._affinity_method("translate", self._data, xoff, yoff, zoff),
             crs=self.crs,
         )
 
-    def rotate(self, angle, origin="center", use_radians=False):
+    def rotate(
+        self, angle, origin="center", use_radians: bool = False
+    ) -> GeometryArray:
         return GeometryArray(
             self._affinity_method(
                 "rotate", self._data, angle, origin=origin, use_radians=use_radians
@@ -1015,7 +1049,13 @@ class GeometryArray(ExtensionArray):
             crs=self.crs,
         )
 
-    def scale(self, xfact=1.0, yfact=1.0, zfact=1.0, origin="center"):
+    def scale(
+        self,
+        xfact: float = 1.0,
+        yfact: float = 1.0,
+        zfact: float = 1.0,
+        origin="center",
+    ) -> GeometryArray:
         return GeometryArray(
             self._affinity_method(
                 "scale", self._data, xfact, yfact, zfact, origin=origin
@@ -1023,7 +1063,13 @@ class GeometryArray(ExtensionArray):
             crs=self.crs,
         )
 
-    def skew(self, xs=0.0, ys=0.0, origin="center", use_radians=False):
+    def skew(
+        self,
+        xs: float = 0.0,
+        ys: float = 0.0,
+        origin="center",
+        use_radians: bool = False,
+    ) -> GeometryArray:
         return GeometryArray(
             self._affinity_method(
                 "skew", self._data, xs, ys, origin=origin, use_radians=use_radians
@@ -1032,7 +1078,7 @@ class GeometryArray(ExtensionArray):
         )
 
     @requires_pyproj
-    def to_crs(self, crs=None, epsg=None):
+    def to_crs(self, crs: Any | None = None, epsg: int | None = None) -> GeometryArray:
         """Transform all geometries to a different coordinate reference system.
 
         Transform all geometries in a GeometryArray to a different coordinate
@@ -1124,7 +1170,7 @@ class GeometryArray(ExtensionArray):
         return GeometryArray(new_data, crs=crs)
 
     @requires_pyproj
-    def estimate_utm_crs(self, datum_name="WGS 84"):
+    def estimate_utm_crs(self, datum_name: str = "WGS 84") -> CRS:
         """Return the estimated UTM CRS based on the bounds of the dataset.
 
         .. versionadded:: 0.9
@@ -1261,9 +1307,6 @@ class GeometryArray(ExtensionArray):
     @property
     def m(self):
         """Return the m coordinate of point geometries in a GeoSeries."""
-        if not SHAPELY_GE_21:
-            raise ImportError("'m' requires shapely>=2.1.")
-
         if (self.geom_type[~self.isna()] == "Point").all():
             empty = self.is_empty
             if empty.any():
@@ -1311,18 +1354,18 @@ class GeometryArray(ExtensionArray):
         return self._data.size
 
     @property
-    def shape(self):
+    def shape(self) -> tuple:
         return (self.size,)
 
     @property
-    def ndim(self):
+    def ndim(self) -> int:
         return len(self.shape)
 
-    def copy(self, *args, **kwargs):
+    def copy(self, *args, **kwargs) -> GeometryArray:
         # still taking args/kwargs for compat with pandas 0.24
         return GeometryArray(self._data.copy(), crs=self._crs)
 
-    def take(self, indices, allow_fill=False, fill_value=None):
+    def take(self, indices, allow_fill: bool = False, fill_value=None) -> GeometryArray:
         from pandas.api.extensions import take
 
         if allow_fill:
@@ -1340,16 +1383,12 @@ class GeometryArray(ExtensionArray):
     def _pad_or_backfill(
         self, method, limit=None, limit_area=None, copy=True, **kwargs
     ):
-        if PANDAS_GE_21 and not PANDAS_GE_22:
-            if limit_area is not None:
-                # limit area not supported, but, but we feed through
-                # so the caller gets the pandas exception
-                kwargs["limit_area"] = limit_area
-        else:
-            kwargs["limit_area"] = limit_area
+        kwargs["limit_area"] = limit_area
         return super()._pad_or_backfill(method=method, limit=limit, copy=copy, **kwargs)
 
-    def fillna(self, value=None, method=None, limit=None, copy=True):
+    def fillna(
+        self, value=None, method=None, limit: int | None = None, copy: bool = True
+    ) -> GeometryArray:
         """
         Fill NA values with geometry (or geometries) or using the specified method.
 
@@ -1357,7 +1396,7 @@ class GeometryArray(ExtensionArray):
         ----------
         value : shapely geometry object or GeometryArray
             If a geometry value is passed it is used to fill all missing values.
-            Alternatively, an GeometryArray 'value' can be given. It's expected
+            Alternatively, a GeometryArray 'value' can be given. It's expected
             that the GeometryArray has the same length as 'self'.
 
         method : {'backfill', 'bfill', 'pad', 'ffill', None}, default None
@@ -1409,7 +1448,7 @@ class GeometryArray(ExtensionArray):
         new_values._data[mask] = value_arr
         return new_values
 
-    def astype(self, dtype, copy=True):
+    def astype(self, dtype, copy: bool = True) -> np.ndarray:
         """
         Cast to a NumPy array with 'dtype'.
 
@@ -1450,14 +1489,14 @@ class GeometryArray(ExtensionArray):
             else:
                 return np.array(self, dtype=dtype, copy=copy)
 
-    def isna(self):
+    def isna(self) -> np.ndarray:
         """Boolean NumPy array indicating if each value is missing."""
         return shapely.is_missing(self._data)
 
     def value_counts(
         self,
         dropna: bool = True,
-    ):
+    ) -> pd.Series:
         """
         Compute a histogram of the counts of non-null values.
 
@@ -1481,11 +1520,11 @@ class GeometryArray(ExtensionArray):
         # Can't use fillna(None) or Index.putmask, as this gets converted back to nan
         # for object dtypes
         result.index = Index(
-            from_wkb(np.where(result.index.isna(), None, result.index))
+            from_wkb(np.where(result.index.isna(), None, result.index), crs=self.crs)
         )
         return result
 
-    def unique(self):
+    def unique(self) -> ExtensionArray:
         """Compute the ExtensionArray of unique values.
 
         Returns
@@ -1501,7 +1540,7 @@ class GeometryArray(ExtensionArray):
     def nbytes(self):
         return self._data.nbytes
 
-    def shift(self, periods=1, fill_value=None):
+    def shift(self, periods: int = 1, fill_value: Any | None = None) -> GeometryArray:
         """
         Shift values by desired number.
 
@@ -1541,7 +1580,7 @@ class GeometryArray(ExtensionArray):
     # -------------------------------------------------------------------------
 
     @classmethod
-    def _from_sequence(cls, scalars, dtype=None, copy=False):
+    def _from_sequence(cls, scalars, dtype=None, copy: bool = False) -> ExtensionArray:
         """
         Construct a new ExtensionArray from a sequence of scalars.
 
@@ -1588,8 +1627,7 @@ class GeometryArray(ExtensionArray):
         # GH 3099
         return from_wkt(strings)
 
-    def _values_for_factorize(self):
-        # type: () -> Tuple[np.ndarray, Any]
+    def _values_for_factorize(self) -> tuple[np.ndarray, Any]:
         """Return an array and missing value suitable for factorization.
 
         Returns
@@ -1608,7 +1646,7 @@ class GeometryArray(ExtensionArray):
         return vals, None
 
     @classmethod
-    def _from_factorized(cls, values, original):
+    def _from_factorized(cls, values: np.ndarray, original: ExtensionArray):
         """
         Reconstruct an ExtensionArray after factorization.
 
@@ -1626,8 +1664,7 @@ class GeometryArray(ExtensionArray):
         """
         return from_wkb(values, crs=original.crs)
 
-    def _values_for_argsort(self):
-        # type: () -> np.ndarray
+    def _values_for_argsort(self) -> np.ndarray:
         """Return values for sorting.
 
         Returns
@@ -1656,7 +1693,7 @@ class GeometryArray(ExtensionArray):
             # process those. The missing values are handled separately by
             # pandas regardless of the values we return here (to sort
             # first/last depending on 'na_position'), the distances for the
-            # empty geometries are substitued below with an appropriate value
+            # empty geometries are replaced below with an appropriate value
             geoms = self.copy()
             indices = np.nonzero(~mask)[0]
             if indices.size:
@@ -1691,7 +1728,7 @@ class GeometryArray(ExtensionArray):
     def argmax(self, skipna: bool = True) -> int:
         raise TypeError("geometries have no minimum or maximum")
 
-    def _formatter(self, boxed=False):
+    def _formatter(self, boxed: bool = False):
         """Return a formatting function for scalar values.
 
         This is used in the default '__repr__'. The returned formatting
@@ -1761,7 +1798,7 @@ class GeometryArray(ExtensionArray):
         data = np.concatenate([ga._data for ga in to_concat])
         return GeometryArray(data, crs=_get_common_crs(to_concat))
 
-    def _reduce(self, name, skipna=True, keepdims=False, **kwargs):
+    def _reduce(self, name: str, skipna: bool = True, keepdims: bool = False, **kwargs):
         # including the base class version here (that raises by default)
         # because this was not yet defined in pandas 0.23
         if name in ("any", "all"):
@@ -1771,7 +1808,7 @@ class GeometryArray(ExtensionArray):
             f"does not support reduction '{name}'"
         )
 
-    def __array__(self, dtype=None, copy=None):
+    def __array__(self, dtype=None, copy=None) -> np.ndarray:
         """Return the data as a numpy array.
 
         This is the numpy array interface.
@@ -1823,7 +1860,7 @@ class GeometryArray(ExtensionArray):
     def __ne__(self, other):
         return self._binop(other, operator.ne)
 
-    def __contains__(self, item):
+    def __contains__(self, item) -> bool:
         """Return for `item in self`."""
         if isna(item):
             if (
@@ -1837,7 +1874,7 @@ class GeometryArray(ExtensionArray):
         return (self == item).any()
 
 
-def _get_common_crs(arr_seq):
+def _get_common_crs(arr_seq) -> CRS:
     # mask out all None arrays with no crs (most likely auto generated by pandas
     # from concat with missing column)
     arr_seq = [ga for ga in arr_seq if not (ga.isna().all() and ga.crs is None)]
@@ -1869,7 +1906,7 @@ def _get_common_crs(arr_seq):
     )
 
 
-def transform(data, func):
+def transform(data, func) -> np.ndarray:
     has_z = shapely.has_z(data)
 
     result = np.empty_like(data)
