@@ -170,6 +170,9 @@ class GeoDataFrame(GeoPandasBase, DataFrame):
 
         super().__init__(data, *args, **kwargs)
 
+        if isinstance(data, DataFrame) and data.attrs:
+            self.attrs = data.attrs
+
         # set_geometry ensures the geometry data have the proper dtype,
         # but is not called if `geometry=None` ('geometry' column present
         # in the data), so therefore need to ensure it here manually
@@ -446,6 +449,8 @@ class GeoDataFrame(GeoPandasBase, DataFrame):
                 )
             if drop:
                 del frame[col]
+                frame.__class__ = GeoDataFrame
+                # revert the casting done in __delitem__, keep gdf
                 warnings.warn(
                     given_colname_drop_msg,
                     category=FutureWarning,
@@ -610,7 +615,7 @@ class GeoDataFrame(GeoPandasBase, DataFrame):
                     "This unsafe behavior will be deprecated in future versions. "
                     "Use GeoDataFrame.set_crs method instead",
                     stacklevel=2,
-                    category=DeprecationWarning,
+                    category=FutureWarning,
                 )
             self.geometry.values.crs = value
         else:
@@ -908,7 +913,7 @@ class GeoDataFrame(GeoPandasBase, DataFrame):
         cls, table, geometry: str | None = None, to_pandas_kwargs: dict | None = None
     ):
         """
-        Construct a GeoDataFrame from a Arrow table object based on GeoArrow
+        Construct a GeoDataFrame from an Arrow table object based on GeoArrow
         extension types.
 
         See https://geoarrow.org/ for details on the GeoArrow specification.
@@ -1124,14 +1129,8 @@ individually so that features may have different properties
         if na not in ["null", "drop", "keep"]:
             raise ValueError(f"Unknown na method {na}")
 
-        if self._geometry_column_name not in self:
-            raise AttributeError(
-                "No geometry data set (expected in column "
-                f"'{self._geometry_column_name}')."
-            )
-
         ids = np.asarray(self.index)
-        geometries = np.asarray(self[self._geometry_column_name])
+        geometries = np.asarray(self.geometry)
 
         if not self.columns.is_unique:
             raise ValueError("GeoDataFrame cannot contain duplicated column names.")
@@ -1502,9 +1501,9 @@ default 'snappy'
         compression : {'zstd', 'lz4', 'uncompressed'}, optional
             Name of the compression to use. Use ``"uncompressed"`` for no
             compression. By default uses LZ4 if available, otherwise uncompressed.
-        schema_version : {'0.1.0', '0.4.0', '1.0.0', None}
+        schema_version : {'0.1.0', '0.4.0', '1.0.0', '1.1.0' None}
             GeoParquet specification version; if not provided will default to
-            latest supported version.
+            latest supported stable version (1.0.0).
         kwargs
             Additional keyword arguments passed to
             :func:`pyarrow.feather.write_feather`.
@@ -1914,6 +1913,12 @@ default 'snappy'
             else:
                 result.__class__ = DataFrame
         return result
+
+    def __delitem__(self, key) -> None:
+        """If the last geometry column is removed, downcast to a dataframe."""
+        super().__delitem__(key)
+        if (self.dtypes == "geometry").sum() == 0:
+            self.__class__ = DataFrame
 
     def _persist_old_default_geometry_colname(self) -> None:
         """Persist the default geometry column name of 'geometry' temporarily for
@@ -2503,6 +2508,24 @@ default 'snappy'
             You can check the valid values in left_df or right_df as
             ``left_df.sindex.valid_query_predicates`` or
             ``right_df.sindex.valid_query_predicates``
+
+            Available predicates include:
+
+            * ``'intersects'``: True if geometries intersect (boundaries and interiors)
+            * ``'within'``: True if left geometry is completely within right geometry
+            * ``'contains'``: True if left geometry completely contains right geometry
+            * ``'contains_properly'``: True if left geometry contains right geometry
+              and their boundaries do not touch
+            * ``'overlaps'``: True if geometries overlap but neither contains the other
+            * ``'crosses':`` True if geometries cross (interiors intersect but neither
+              contains the other, with intersection dimension less than max dimension)
+            * ``'touches'``: True if geometries touch at boundaries but interiors don't
+            * ``'covers'``: True if left geometry covers right geometry (every point of
+              right is a point of left)
+            * ``'covered_by'``: True if left geometry is covered by right geometry
+            * ``'dwithin'``: True if geometries are within specified distance (requires
+              distance parameter)
+
         lsuffix : string, default 'left'
             Suffix to apply to overlapping column names (left GeoDataFrame).
         rsuffix : string, default 'right'
