@@ -18,7 +18,7 @@ from shapely.geometry import Point, Polygon, box, mapping
 
 import geopandas
 from geopandas import GeoDataFrame, GeoSeries, points_from_xy, read_file
-from geopandas._compat import HAS_PYPROJ, PANDAS_GE_30
+from geopandas._compat import HAS_PYPROJ, PANDAS_GE_30, PANDAS_INFER_STR
 from geopandas.io.file import _EXTENSION_TO_DRIVER, _detect_driver
 
 import pytest
@@ -34,9 +34,17 @@ try:
     PYOGRIO_GE_090 = Version(Version(pyogrio.__version__).base_version) >= Version(
         "0.9.0"
     )
+    PYOGRIO_GE_012 = Version(Version(pyogrio.__version__).base_version) >= Version(
+        "0.12.0"
+    )
+    PYOGRIO_GE_0121 = Version(Version(pyogrio.__version__).base_version) >= Version(
+        "0.12.1"
+    )
 except ImportError:
     pyogrio = False
     PYOGRIO_GE_090 = False
+    PYOGRIO_GE_012 = False
+    PYOGRIO_GE_0121 = False
 
 
 try:
@@ -197,7 +205,7 @@ def test_to_file_bool(tmpdir, driver, ext, engine):
 
 TEST_DATE = datetime.datetime(2021, 11, 21, 1, 7, 43, 17500)
 # from pandas 2.0, utc equality checks less stringent, forward compat with zoneinfo
-utc = datetime.timezone.utc
+utc = datetime.UTC
 eastern = zoneinfo.ZoneInfo("America/New_York")
 test_date_eastern = TEST_DATE.replace(tzinfo=eastern)
 datetime_type_tests = (TEST_DATE, test_date_eastern)
@@ -274,7 +282,7 @@ def test_read_file_datetime_invalid(tmpdir, ext, engine):
         assert is_datetime64_any_dtype(res["date"])
         assert pd.isna(res["date"].iloc[-1])
     else:
-        assert res["date"].dtype == "str" if PANDAS_GE_30 else object
+        assert res["date"].dtype == "str" if PANDAS_INFER_STR else object
         assert isinstance(res["date"].iloc[-1], str)
 
 
@@ -283,13 +291,19 @@ def test_read_file_datetime_out_of_bounds_ns(tmpdir, ext, engine):
     # https://github.com/geopandas/geopandas/issues/2502
     date_str = "9999-12-31T00:00:00"  # valid to GDAL, not to [ns] format
     tempfilename = write_invalid_date_file(date_str, tmpdir, ext, engine)
-    res = read_file(tempfilename, engine=engine)
+    if engine == "pyogrio" and PYOGRIO_GE_012 and not PANDAS_GE_30:
+        with pytest.warns(
+            UserWarning, match="Error parsing datetimes, original strings are returned"
+        ):
+            res = read_file(tempfilename, engine=engine)
+    else:
+        res = read_file(tempfilename, engine=engine)
     if PANDAS_GE_30:
         assert res["date"].dtype == "datetime64[ms]"
         assert res["date"].iloc[-1] == pd.Timestamp("9999-12-31 00:00:00")
     else:
         # Pandas invalid datetimes are read in as object dtype (strings)
-        assert res["date"].dtype == "object"
+        assert res["date"].dtype == "str" if PANDAS_INFER_STR else "object"
         assert isinstance(res["date"].iloc[0], str)
 
 
@@ -494,7 +508,7 @@ def test_to_file_with_duplicate_columns(tmpdir, engine):
     df = GeoDataFrame(data=[[1, 2, 3]], columns=["a", "b", "a"], geometry=[Point(1, 1)])
     tempfilename = os.path.join(str(tmpdir), "duplicate.shp")
     with pytest.raises(
-        ValueError, match="GeoDataFrame cannot contain duplicated column names."
+        ValueError, match="GeoDataFrame cannot contain duplicated column names"
     ):
         df.to_file(tempfilename, engine=engine)
 
@@ -593,13 +607,27 @@ def test_read_file(engine, nybb_filename):
         # url to zip file
         "https://raw.githubusercontent.com/geopandas/geopandas/"
         "main/geopandas/tests/data/nybb_16a.zip",
-        # url to zipfile without extension
-        "https://geonode.goosocean.org/download/480",
         # url to web service
         "https://demo.pygeoapi.io/stable/collections/obs/items",
     ],
 )
 def test_read_file_url(engine, url):
+    gdf = read_file(url, engine=engine)
+    assert isinstance(gdf, geopandas.GeoDataFrame)
+
+
+@pytest.mark.web
+@pytest.mark.parametrize(
+    "url",
+    [
+        # url to zipfile without extension
+        "https://geonode.goosocean.org/download/480",
+    ],
+)
+# Marked strict=False as state is not stable see
+# #3584, #3585, #3644
+@pytest.mark.xfail(reason="SSL certificate issue", strict=False)
+def test_read_file_url_flaky(engine, url):
     gdf = read_file(url, engine=engine)
     assert isinstance(gdf, geopandas.GeoDataFrame)
 
@@ -863,7 +891,7 @@ def test_read_file_filtered__rows_bbox(df_nybb, engine, nybb_filename):
         # TODO: support negative rows in pyogrio
         with pytest.raises(
             ValueError,
-            match="'skip_features' must be between 0 and 1|Negative slice start",
+            match=r"'skip_features' must be between 0 and 1|Negative slice start",
         ):
             filtered_df = read_file(
                 nybb_filename, bbox=bbox, rows=slice(-1, None), engine=engine
@@ -980,7 +1008,13 @@ def test_read_file__columns(engine, naturalearth_lowres):
     gdf = geopandas.read_file(
         naturalearth_lowres, columns=["name", "pop_est"], engine=engine
     )
-    assert gdf.columns.tolist() == ["name", "pop_est", "geometry"]
+
+    expected = (
+        ["pop_est", "name", "geometry"]
+        if engine == "pyogrio" and PYOGRIO_GE_0121
+        else ["name", "pop_est", "geometry"]
+    )
+    assert gdf.columns.tolist() == expected
 
 
 def test_read_file__columns_empty(engine, naturalearth_lowres):
@@ -1009,7 +1043,13 @@ def test_read_file__include_fields(engine, naturalearth_lowres):
     gdf = geopandas.read_file(
         naturalearth_lowres, include_fields=["name", "pop_est"], engine=engine
     )
-    assert gdf.columns.tolist() == ["name", "pop_est", "geometry"]
+
+    expected = (
+        ["pop_est", "name", "geometry"]
+        if engine == "pyogrio" and PYOGRIO_GE_0121
+        else ["name", "pop_est", "geometry"]
+    )
+    assert gdf.columns.tolist() == expected
 
 
 @pytest.mark.skipif(not FIONA_GE_19, reason="columns requires fiona 1.9+")
@@ -1328,7 +1368,7 @@ def test_write_index_to_file(tmpdir, df_points, driver, ext, engine):
     df.index += 1
     do_checks(df, index_is_used=False)
 
-    # index is a Int64Index regular sequence from 1
+    # index is an Int64Index regular sequence from 1
     df_p.index = list(range(1, len(df) + 1))
     df = GeoDataFrame(df_p["value1"], geometry=df_p.geometry)
     do_checks(df, index_is_used=False)
@@ -1511,6 +1551,40 @@ def test_error_engine_unavailable_fiona(tmp_path, df_points, file_path):
         geopandas.read_file(file_path, engine="fiona")
 
     with pytest.raises(ImportError, match="the 'to_file' method requires"):
+        df_points.to_file(tmp_path / "test.gpkg", engine="fiona")
+
+
+def test_error_monkeypatch_engine_unavailable_pyogrio(
+    monkeypatch, tmp_path, df_points, file_path
+) -> None:
+    # monkeypatch to make pyogrio unimportable
+    monkeypatch.setattr(geopandas.io.file, "_import_pyogrio", lambda: None)
+    monkeypatch.setattr(geopandas.io.file, "pyogrio", None)
+    monkeypatch.setattr(
+        geopandas.io.file, "pyogrio_import_error", "No module named 'pyogrio'"
+    )
+
+    with pytest.raises(ImportError, match="No module named 'pyogrio'"):
+        geopandas.read_file(file_path, engine="pyogrio")
+
+    with pytest.raises(ImportError, match="No module named 'pyogrio'"):
+        df_points.to_file(tmp_path / "test.gpkg", engine="pyogrio")
+
+
+def test_error_monkeypatch_engine_unavailable_fiona(
+    monkeypatch, tmp_path, df_points, file_path
+) -> None:
+    # monkeypatch to make fiona unimportable
+    monkeypatch.setattr(geopandas.io.file, "_import_fiona", lambda: None)
+    monkeypatch.setattr(geopandas.io.file, "fiona", None)
+    monkeypatch.setattr(
+        geopandas.io.file, "fiona_import_error", "No module named 'fiona'"
+    )
+
+    with pytest.raises(ImportError, match="No module named 'fiona'"):
+        geopandas.read_file(file_path, engine="fiona")
+
+    with pytest.raises(ImportError, match="No module named 'fiona'"):
         df_points.to_file(tmp_path / "test.gpkg", engine="fiona")
 
 
