@@ -97,7 +97,7 @@ class TestSeries:
     def test_align_mixed(self):
         a1 = self.a1
         s2 = pd.Series([1, 2], index=["B", "C"])
-        res1, res2 = a1.align(s2)
+        _res1, res2 = a1.align(s2)
 
         exp2 = pd.Series([np.nan, 1, 2], index=["A", "B", "C"])
         assert_series_equal(res2, exp2)
@@ -159,7 +159,6 @@ class TestSeries:
             self.a1.geom_equals_exact(self.a2, 0.001, align=False), [False, False]
         )
 
-    @pytest.mark.skipif(not compat.SHAPELY_GE_21, reason="requires Shapely>=2.1")
     def test_geom_equals_identical(self):
         assert np.all(self.g1.geom_equals_identical(self.g1))
         assert_array_equal(self.g1.geom_equals_identical(self.sq), [False, True])
@@ -538,7 +537,7 @@ def test_geoseries_crs():
 def test_geoseries_override_existing_crs_warning():
     gs = GeoSeries(crs="epsg:4326")
     with pytest.warns(
-        DeprecationWarning,
+        FutureWarning,
         match="Overriding the CRS of a GeoSeries that already has CRS",
     ):
         gs.crs = "epsg:2100"
@@ -628,7 +627,7 @@ class TestConstructor:
             np.array([], dtype="float64"),
         ]
 
-        if compat.PANDAS_GE_30 and pd.options.future.infer_string:
+        if compat.PANDAS_INFER_STR:
             # in pandas >=3 future string, str is not converted to object
             # so is non geom compatible
             non_geom_compat_dtypes.append(np.array([], dtype="str"))
