@@ -40,7 +40,23 @@ def sjoin(
         You can check the valid values in left_df or right_df as
         ``left_df.sindex.valid_query_predicates`` or
         ``right_df.sindex.valid_query_predicates``
-        Replaces deprecated ``op`` parameter.
+
+        Available predicates include:
+
+        * ``'intersects'``: True if geometries intersect (boundaries and interiors)
+        * ``'within'``: True if left geometry is completely within right geometry
+        * ``'contains'``: True if left geometry completely contains right geometry
+        * ``'contains_properly'``: True if left geometry contains right geometry
+          and their boundaries do not touch
+        * ``'overlaps'``: True if geometries overlap but neither contains the other
+        * ``'crosses':`` True if geometries cross (interiors intersect but neither
+          contains the other, with intersection dimension less than max dimension)
+        * ``'touches'``: True if geometries touch at boundaries but interiors don't
+        * ``'covers'``: True if left geometry covers right geometry (every point of
+          right is a point of left)
+        * ``'covered_by'``: True if left geometry is covered by right geometry
+        * ``'dwithin'``: True if geometries are within specified distance (requires
+          distance parameter)
     lsuffix : string, default 'left'
         Suffix to apply to overlapping column names (left GeoDataFrame).
     rsuffix : string, default 'right'
@@ -345,7 +361,7 @@ def _process_column_names_with_suffix(
 
 def _restore_index(joined, index_names, index_names_original):
     """
-    Set back the the original index columns, and restoring their name as `None`
+    Set back the original index columns, and restoring their name as `None`
     if they didn't have a name originally.
     """
     if PANDAS_GE_30:
